feat: add PrimitiveMesh utils class
